fix(bkb) drag-scroll was dumping spill-over distance >
reducing effectively the attenable scroll-speed.
From another POV, scroll-buffer's input-speed division was getting rid
of the remainder.

The effect was a subtle jerk with sensible defaults (SCROLL_DPI: 100)
but in higher DPIs or
with *maccel* [1] enabled and redirecting to drag-scroll,
all expected addionall velocity, dissipates.

NOTE: to enable *maccel*, you need to reverse the lines in
`pointing_device_task_kb()` so that `pointing_device_task_user()`
is invoked first.

WARNING: there might be an issue when drag-scroll DPI is 100[2].

[1]: https://github.com/finrod09/qmk_userspace_features/tree/maccel-assym_s_curve/maccel
[2]: https://discord.com/channels/681309835135811648/1203417572742135808/1209118208997466142